Update jquery.loxiatable-2.js
add option "buttons".
in this case, I'll add batch operation button at the navigation bar. If it's a 'sync' batch operation button, the button will be surrounded with a form, otherwise it's a simple button with class attribute 'batchOperation'.

Now, if user want to add events on this button, they should add it manually.
In the future, I'll add event on this button in this file.